+++ src/silicone/database_crunchers/rms_closest.py
@@ -77,20 +77,28 @@ class RMSClosest(_DatabaseCruncher):
         iamdf_follower = self._get_iamdf_section(variable_follower)
         data_follower_time_col = iamdf_follower.time_col
         iamdf_lead = self._db.filter(variable=variable_leaders)
+
         if not weighting:
             weighting = {variab: 1 for variab in variable_leaders}
+
         if any(var not in weighting.keys() for var in variable_leaders):
             raise ValueError("Weighting does not include all lead variables.")
-        iamdf_lead, iamdf_follower = _filter_for_overlap(
+
+        iamdf_lead_data, iamdf_follower_data = _filter_for_overlap(
             iamdf_lead,
             iamdf_follower,
             ["scenario", "model", data_follower_time_col],
             variable_leaders,
         )
 
+        iamdf_lead_ts = pyam.IamDataFrame(iamdf_lead_data).timeseries()
+        iamdf_follower_ts = pyam.IamDataFrame(iamdf_follower_data).timeseries()
+
         leader_var_unit = {
             var["variable"]: var["unit"]
-            for _, var in iamdf_lead[["variable", "unit"]].drop_duplicates().iterrows()
+            for _, var in iamdf_lead_data[["variable", "unit"]]
+            .drop_duplicates()
+            .iterrows()
         }
 
         def filler(in_iamdf):
@@ -139,70 +147,94 @@ class RMSClosest(_DatabaseCruncher):
                     )
                 )
 
-            lead_var_timeseries = lead_var.timeseries()
-            iamdf_lead_timeseries = iamdf_lead.pivot(
-                index=[
-                    col
-                    for col in iamdf_lead.columns
-                    if col not in [data_follower_time_col, "value"]
-                ],
-                columns=data_follower_time_col,
-                values="value",
+            (
+                iamdf_lead_ts_here,
+                lead_var_ts_here,
+                common_cols,
+            ) = _get_common_timeseries_and_cols(
+                iamdf_lead_ts.copy(), lead_var.timeseries()
             )
-            common_cols = [
-                col
-                for col in lead_var_timeseries.columns
-                if col in iamdf_lead_timeseries.columns
-            ]
-            if not common_cols:
-                raise ValueError(
-                    "No time series overlap between the original and unfilled data"
-                )
+            iamdf_lead_ts_here = iamdf_lead_ts_here[common_cols]
+            lead_var_ts_here = lead_var_ts_here[common_cols]
 
-            lead_var_timeseries = lead_var_timeseries.loc[:, common_cols]
-            iamdf_lead_timeseries = iamdf_lead_timeseries.loc[:, common_cols].dropna(
-                axis=0
+            rms = _calculate_rms(
+                iamdf_lead_ts_here,
+                lead_var_ts_here,
+                weighting,
             )
 
-            output_ts_list = []
-            for _, (model, scenario) in (
-                lead_var.data[["model", "scenario"]].drop_duplicates().iterrows()
-            ):
-                lead_var_mod_scen = lead_var_timeseries[
-                    (lead_var_timeseries.index.get_level_values("model") == model)
-                    & (
-                        lead_var_timeseries.index.get_level_values("scenario")
-                        == scenario
-                    )
-                ]
-                if len(lead_var_mod_scen) != len(variable_leaders):
-                    raise ValueError(
-                        "Insufficient variables are found to infill model {}, scenario "
-                        "{}. Only found {}.".format(model, scenario, lead_var_mod_scen)
-                    )
-                closest_model, closest_scenario = _select_closest(
-                    iamdf_lead_timeseries,
-                    lead_var_mod_scen,
-                    weighting,
-                    variable_leaders,
-                )
+            out = _combine_rms_and_database(
+                lead_var_ts_here,
+                iamdf_follower_ts,
+                rms,
+                [variable_follower],
+            )
 
-                # Filter to find the matching follow data for the same model, scenario
-                # and region
-                tmp = iamdf_follower.loc[
-                    (iamdf_follower.model == closest_model)
-                    & (iamdf_follower.scenario == closest_scenario)
-                ].copy()
-
-                # Update the model and scenario to match the elements of the input.
-                tmp.loc[:, "model"] = model
-                tmp.loc[:, "scenario"] = scenario
-                for col in in_iamdf.extra_cols:
-                    tmp[col] = lead_var_mod_scen.index.get_level_values(col).tolist()[0]
-                output_ts_list.append(tmp)
-            return pyam.concat(output_ts_list)
+            return out
 
         return filler
+
+    def infill_multiple(
+        self, to_infill, variable_followers, variable_leaders, weighting=None
+    ):
+        """
+        Infill multiple variables simultaneously
+
+        This can be much faster as the RMS between the lead and follower
+        scenarios only needs to be calculated once.
+
+        Parameters
+        ----------
+        to_infill : :class:`pyam.IamDataFrame`
+            The timeseries to infill
+
+        variable_followers : list[str]
+            The variables for which we want to infill timeseries (e.g.
+            ``["Emissions|C5F12", "Emissions|C4F10"]``).
+
+        variable_leaders : list[str]
+            The variable we want to use in order to infer timeseries of
+            ``variable_follower`` (e.g. ``["Emissions|CO2"]``). This may contain
+            multiple elements.
+
+        weighting : dict{str: float}
+            When used with multiple lead variables, this weighting factor controls the
+            relative importance of different variables for determining closeness. E.g.
+            if wanting to compare both CO2 and CH4 emissions reported in mass
+            units but weighted by the AR5 GWP100 metric, this would be
+            {"Emissions|CO2": 1, "Emissions|CH4": 28}.
+
+        Returns
+        -------
+        :class:`pyam.IamDataFrame`
+            Infilled timeseries
+
+        Raises
+        ------
+        ValueError
+            There is no data for ``variable_leaders`` or ``variable_follower`` in the
+            database.
+        """
+        self._check_followers_and_leaders_in_db(variable_followers, variable_leaders)
+        db_lead = self._db.filter(variable=variable_leaders)
+        to_infill_lead = to_infill.filter(variable=variable_leaders)
+
+        db_lead_ts, to_infill_lead_ts, common_cols = _get_common_timeseries_and_cols(
+            db_lead.timeseries(), to_infill_lead.timeseries()
+        )
+
+        rms = _calculate_rms(
+            db_lead_ts[common_cols], to_infill_lead_ts[common_cols], weighting
+        )
+
+        db_timeseries = self._db.filter(variable=variable_followers).timeseries()
+        db_timeseries = db_timeseries[common_cols]
+
+        out = _combine_rms_and_database(
+            to_infill_lead_ts, db_timeseries, rms, variable_followers
+        )
+
+        return out
 
     def _check_iamdf_lead(self, variable_leaders):
         if not all([v in self._db.variable for v in variable_leaders]):
@@ -226,55 +258,134 @@ class RMSClosest(_DatabaseCruncher):
 
         return iamdf_section
 
+    def _check_followers_and_leaders_in_db(self, variable_followers, variable_leaders):
+        if not all([v in self._db.variable for v in variable_leaders]):
+            error_msg = "No data for `variable_leaders` ({}) in database".format(
+                variable_leaders
+            )
+            raise ValueError(error_msg)
 
-def _select_closest(to_search_df, target_df, weighting, variable_leaders):
-    """
-    Find model/scenario combo in ``to_search_df`` that is closest to that of the target.
+        if not all([v in self._db.variable for v in variable_followers]):
+            error_msg = "No data for `variable_followers` ({}) in database".format(
+                variable_followers
+            )
+            raise ValueError(error_msg)
 
-    Here, 'closest' is in the root-mean squared sense. In the event that multiple model/
-    scenarios are equally close, returns first row.
 
-    Parameters
-    ----------
-    to_search_df : :obj:`pd.DataFrame`
-        The model/scenario combos to search for the closest case. A timeseries.
+def _get_common_timeseries_and_cols(db_lead, to_infill_lead):
+    db_lead.index = db_lead.index.rename(
+        ["model_db", "scenario_db"], level=["model", "scenario"]
+    )
 
-    target_df : :obj:`pd.DataFrame`
-        The data to which we want to be close. A timeseries.
+    to_infill_lead.index = to_infill_lead.index.rename(
+        ["model_lead", "scenario_lead"], level=["model", "scenario"]
+    )
 
-    weighting : map{str: float}
-        Maps the variable name onto the weighting for that variable.
+    common_cols = _get_common_cols(db_lead, to_infill_lead)
 
-    Returns
-    -------
-    dict
-        Index of the closest timeseries.
-    """
+    return db_lead, to_infill_lead, common_cols
 
-    rms = pd.Series(0, index=to_search_df.index, dtype=np.float64)
-    for var in variable_leaders:
-        target_for_var = target_df[
-            target_df.index.get_level_values("variable") == var
-        ].squeeze()
-        rms = rms.add(
-            (
-                to_search_df[
-                    to_search_df.index.get_level_values("variable") == var
-                ].subtract(target_for_var, axis=1)
-                ** 2
-            ).mean(axis=1)
-            ** 0.5
-            * weighting[var],
-            fill_value=0,
-        )
-    rmssums = rms.groupby(level=["model", "scenario"], sort=False).sum()
-    return rmssums.idxmin()
+
+def _calculate_rms(db_lead_ts, to_infill_lead_ts, weighting):
+    # check variables first
+    ms_lead = ["model_lead", "scenario_lead"]
+
+    required_vars = set(db_lead_ts.index.get_level_values("variable").unique())
+    for (model, scenario), msdf in to_infill_lead_ts.groupby(ms_lead):
+        msdf_vars = set(msdf.index.get_level_values("variable").unique())
+        if msdf_vars != required_vars:
+            raise ValueError(
+                f"Insufficient variables are found to infill model {model}, scenario "
+                f"{scenario}. Only found {msdf}."
+            )
+
+    # remove any database timeseries which have nans
+    db_lead_ts = db_lead_ts.dropna()
+
+    # align
+    db_lead_ts, to_infill_lead_ts = db_lead_ts.align(to_infill_lead_ts)
+
+    rms = ((db_lead_ts - to_infill_lead_ts) ** 2).mean(axis=1) ** 0.5
+
+    if weighting is not None:
+        weighting = pd.Series(weighting)
+        weighting.index.names = ["variable"]
+
+        rms = rms.multiply(weighting)
+
+    rms = rms.groupby(["model_lead", "scenario_lead", "model_db", "scenario_db"]).sum()
+
+    return rms
+
+
+def _combine_rms_and_database(
+    to_infill_lead_ts, db_timeseries, rms, variable_followers
+):
+    out = []
+    for (model, scenario), to_infill_lead_ts_mod_scen in to_infill_lead_ts.groupby(
+        ["model_lead", "scenario_lead"]
+    ):
+        variable_followers_h = set(variable_followers)
+
+        rms_mod_scen = rms.loc[
+            (rms.index.get_level_values("model_lead") == model)
+            & (rms.index.get_level_values("scenario_lead") == scenario),
+            :,
+        ]
+        for (model_db, scenario_db), _ in (
+            rms_mod_scen[(model, scenario)].sort_values().iteritems()
+        ):
+            infill_timeseries = db_timeseries.loc[
+                (db_timeseries.index.get_level_values("model") == model_db)
+                & (db_timeseries.index.get_level_values("scenario") == scenario_db)
+                & (
+                    db_timeseries.index.get_level_values("variable").isin(
+                        variable_followers_h
+                    )
+                ),
+                :,
+            ].copy()
+
+            variable_followers_h = variable_followers_h - set(
+                infill_timeseries.index.get_level_values("variable")
+            )
+
+            infill_timeseries = infill_timeseries.reset_index()
+            infill_timeseries.loc[:, "model"] = model
+            infill_timeseries.loc[:, "scenario"] = scenario
+
+            for idx_level in to_infill_lead_ts.index.names:
+                if idx_level in infill_timeseries or idx_level in [
+                    "model_lead",
+                    "scenario_lead",
+                ]:
+                    continue
+
+                val_to_use = to_infill_lead_ts_mod_scen.index.get_level_values(
+                    idx_level
+                ).unique()
+                if len(val_to_use) != 1:
+                    raise AssertionError(
+                        f"Ambiguous value to use for column {idx_level}, found {val_to_use}"
+                    )
+
+                infill_timeseries[idx_level] = val_to_use[0]
+
+            out.append(infill_timeseries)
+
+            if not variable_followers_h:
+                break
+
+    out = pyam.IamDataFrame(pd.concat(out))
+
+    return out
 
 
 def _filter_for_overlap(df1, df2, cols, leaders):
     """
     Returns overlapping model/scenario combinations in the two input dataframes, which
     must have the same columns.
+
     Parameters
     ----------
     df1 : :obj:`pd.DataFrame`
@@ -287,6 +398,7 @@ def _filter_for_overlap(df1, df2, cols, leaders):
     leaders : list[str]
         List of lead variables that must be found in all acceptable model/scenarios
         combinations.
+
     Returns
     -------
     (:obj:`pd.DataFrame`, :obj:`pd.DataFrame`)
@@ -295,8 +407,8 @@ def _filter_for_overlap(df1, df2, cols, leaders):
     """
     lead_data = df1.data.set_index(cols)
     follow_data = df2.data.set_index(cols)
-    # We only want to select model/scenario cases where we have data for all leaders
 
+    # We only want to select model/scenario cases where we have data for all leaders
     shared_indices = lead_data.index[
         lead_data.index.isin(follow_data.index)
     ].value_counts()
@@ -307,4 +419,15 @@ def _filter_for_overlap(df1, df2, cols, leaders):
 
     lead_data = lead_data.loc[shared_indices]
     follow_data = follow_data.loc[shared_indices]
+
     return lead_data.reset_index(), follow_data.reset_index()
+
+
+def _get_common_cols(lead, base):
+    common_cols = lead.columns.intersection(base.columns)
+    if common_cols.empty:
+        raise ValueError(
+            "No time series overlap between the original and unfilled data"
+        )
+
+    return common_cols